Snowflake_connection_with_private_key/requirements.txt
cryptography
python-dotenv
apscheduler
snowflake-connector